Fixed two defects and added the new file MediaCounterStatus.java that was missing from the previous commit. The two defects were that the state (and color) would not correctly update when the count was incremented, and that extraneous episodes could be added to the database when not allowed.

+++ src/com/example/MediaCounterApp/Activity/MediaCounterActivity.java
@@ -241,15 +241,16 @@ public class MediaCounterActivity extends Activity
             int pos = lv.getPositionForView(ll);
             MediaData md = (MediaData) lv.getAdapter().getItem(pos);
 
-            md.adjustCount(increment);
-
-            if (increment)
+            if (md.adjustCount(increment))
             {
-                db.addEpisode(md.getMediaName());
-            }
-            else
-            {
-                db.deleteEpisode(md.getMediaName());
+                if (increment)
+                {
+                    db.addEpisode(md.getMediaName());
+                }
+                else
+                {
+                    db.deleteEpisode(md.getMediaName());
+                }
             }
 
             if (!increment && (md.getCount() < 0))
